test-requirements.txt
juju<3.1
pytest-operator<1.0
aiohttp<3.8
asyncio<3.5
beautifulsoup4<4.11
PyYAML==6.0
